+++ frontend/src/Settings/Users/index.js
@@ -1,0 +1,3 @@
+import UsersConnector from './UsersConnector';
+
+export default UsersConnector;